% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakers.R
\docType{data}
\name{speakers}
\alias{speakers}
\title{Number of speakers for 370 languages}
\format{A data frame with columns:
\describe{
  \item{language}{Three letter language code.}
  \item{speakers}{Number of speakers.}
  \item{name}{Full name of language.}
  \item{iso6391}{ISO 639-1 codes. See more at
    \url{http://en.wikipedia.org/wiki/ISO_639}.}
  \item{iso6392}{ISO 639-2T codes. See more at
    \url{http://en.wikipedia.org/wiki/ISO_639}.}
}}
\usage{
speakers
}
\description{
This is a superset of all languages detected by franc. Numbers were
collected by Titus Wormer. To quote him: \emph{Painstakingly crawled by
hand from OHCHR, the numbers are (in some cases, very) rough estimates
or out-of-date.}.
}
\keyword{datasets}

